requirements.txt
# Core UI & Web
streamlit>=1.28
requests>=2.31

# AI & Language Models
openai>=1.0

# Audio & Voice Synthesis
elevenlabs>=0.2.0

# Image & Video Processing
pillow>=10.0
numpy>=1.24
opencv-python>=4.8.0
ffmpeg-python>=0.2.1

# Framework & Data
pydantic>=2.0
python-dotenv>=1.0
flask>=3.0
pandas>=2.0

# YouTube & Google APIs
google-auth-oauthlib>=1.0
google-auth-httplib2>=0.2
google-api-python-client>=2.80
google-auth>=2.25

# Optional: ComfyUI Integration (if using local ComfyUI)
# Install ComfyUI separately: https://github.com/comfyanonymous/ComfyUI

# System Dependencies (Must Install Separately):
# - FFmpeg: https://ffmpeg.org/download.html
#   Windows: choco install ffmpeg
#   Mac: brew install ffmpeg
#   Linux: sudo apt-get install ffmpeg

# Development (Optional)
# pytest>=7.0
# black>=23.0
# flake8>=6.0

